add edit and detail

+++ assets/js/dashboard.js
@@ -1,39 +1,40 @@
 const urlMovies = "https://5e8ecf49fe7f2a00165ee9ff.mockapi.io/movies"
 const urlUsers = "https://5e8ecf49fe7f2a00165ee9ff.mockapi.io/users"
 const movies = [];
-const tampilan = document.getElementById("main");
+const tampilan = document.getElementById("mainDetail");
+const detail = document.getElementById("aboutMovie");
 
 let check = localStorage.getItem('isLogin');
 
-if(check != 'true'){
+if (check != 'true') {
     window.location.href = `${window.origin}/index.html`
 }
 
 const getLocalStorage = () => {
-    return localStorage.getItem("userData")===null
-    ? []
-    : JSON.parse(localStorage.getItem("userData"));
+    return localStorage.getItem("userData") === null
+        ? []
+        : JSON.parse(localStorage.getItem("userData"));
 }
 
-const tampilUsername = (user = getLocalStorage()) =>{
+const tampilUsername = (user = getLocalStorage()) => {
     let showUsername = document.getElementById("username");
     console.log(user);
-    
+
     showUsername.innerHTML = ` <a class="dropdown-item" href="#">${user.username}</a>
     <a class="dropdown-item" href="#" onclick="logout(event)">Logout</a>`
 }
 
 const getUserId = (userId = getLocalStorage()) => {
-    return userId.id;   
+    return userId.id;
 }
-const getMovies = async() => {
+const getMovies = async () => {
     const response = await fetch(urlMovies);
     const result = await response.json();
     const userId = getUserId();
-    
+
     result.forEach(element => {
-       movies.push(element) 
-       if(userId == element.userId){
+        movies.push(element)
+        if (userId == element.userId) {
             const newDiv = document.createElement("div")
             newDiv.innerHTML = ` <div class="container" id="mainContainer">
                 <div class="col-md-12 card text-center">
@@ -47,9 +48,9 @@ const getMovies = async() => {
                         alt="">
                     <br>
                     <div>
-                        <button type="submit" id="edit-${element.id}" class="edit-button btn btn-dark" onclick = "editMovies(event)">Edit</button>
+                        <button type="submit" id="edit-${element.id}" class="edit-button btn btn-dark" onclick = "editMovies(event)" data-toggle="modal" data-target="#myModalEdit">Edit</button>
                         <button type="submit" id="hapus-${element.id}" class="delete-button btn btn-dark" onclick = "deleteMovies(event)">Hapus</button>
-                        <button type="submit" id="detail-${element.id}" class="delete-button btn btn-dark" onclick = "detailMovies(event)">Detail</button>
+                        <button type="submit" id="detail-${element.id}" class="delete-button btn btn-dark" onclick = "detailMovies(event)" data-toggle="modal" data-target="#myModalDetail">Detail</button>
                     </div>
                     </div>
                     <div class="card-footer text-muted">
@@ -57,43 +58,75 @@ const getMovies = async() => {
                     </div>
                 </div>
                 </div>`
-            tampilan.appendChild(newDiv)    
-       }
-    });   
+            tampilan.appendChild(newDiv)
+        }
+    });
 }
 
-const deleteMovies = async(event)=>{
-    alert("Yakin Untuk Menghapus ?")
-    const id = event.target.id.replace("hapus-", "");
-    console.log(id);
-    
-    const response = await fetch(`${urlMovies}/${id}`,{
-        method : "DELETE",
-    })
-    await response.json();
-    location.reload();
+const deleteMovies = async (event) => {
+    confirm("Yakin Untuk Menghapus ?")
+    if (confirm == true) {
+        const id = event.target.id.replace("hapus-", "");
+        console.log(id);
+
+        const response = await fetch(`${urlMovies}/${id}`, {
+            method: "DELETE",
+        })
+        await response.json();
+        location.reload();
+    }
+
 }
 
-const editMovies = async(event)=>{
-    alert("Yakin Untuk Mengedit")
-    const editRating = prompt("Insert Your New Rating 0-10");
-    const newRating = parseFloat(editRating);
-    if(newRating>10){
-        alert("Rating Maksimal 9.9")
+const editMovies = async (event) => {
+    const id = event.target.id.replace("edit-", "");
+    const response = await fetch(`${urlMovies}/${id}`);
+    const result = await response.json();
+    const newDiv = document.getElementById("editMovie");
+    newDiv.innerHTML = `<div class="form-group">
+                            <input type="email" class="form-control" id="edit-title" placeholder="Title" value = "${result.title}">
+                        </div>
+                        <div class="form-group">
+                            <textarea class="form-control" id="edit-about" rows="5" placeholder="About">${result.about}</textarea>
+                        </div>
+                        <div class="form-group">
+                            <input type="number" class="form-control" id="edit-rating" placeholder="Rating" value = "${result.rating}">
+                        </div>
+                        <button  type="submit" id="update-${result.id}" onclick = "updateMovies(event)" class="btn btn-dark">Update</button>`
+}
+
+const updateMovies = async(event) =>{
+    event.preventDefault();
+    const title = document.getElementById("edit-title").value;
+    const about = document.getElementById("edit-about").value;
+    const rating = document.getElementById("edit-rating").value;
+    const ratingNum = parseFloat(rating);
+    const update = {
+        title,about,rating,
+    };
+    if(ratingNum>=10){
+        alert("Maksimal rating 10")
     }
     else{
-        const id = event.target.id.replace("edit-", "");
+        const id = event.target.id.replace("update-", "")
         const response = await fetch(`${urlMovies}/${id}`,{
             method : "PUT",
             headers : {
                 "Content-Type" : "application/json",
             },
-            body : JSON.stringify({rating: newRating}),
+            body : JSON.stringify(update),
         });
         await response.json();
+        alert("Data Updated !");
         location.reload();
+
     }
     
+    
+    
+    
+
+
 }
 
 const logout = (event) => {
@@ -101,6 +134,14 @@ const logout = (event) => {
     localStorage.setItem('isLogin', false)
     localStorage.removeItem("userData");
     window.location.href = `${window.origin}/index.html`
+}
+
+const detailMovies = async (event) => {
+    const id = event.target.id.replace("detail-", "");
+    const response = await fetch(`${urlMovies}/${id}`);
+    const result = await response.json();
+    const newDiv = document.getElementById("aboutMovie");
+    newDiv.innerHTML = `${result.about}`;
 }
 
 getMovies();
